+++ app/workers/bq_grapher.py
@@ -5,7 +5,7 @@ import os
 from networkx import DiGraph, write_gpickle
 
 from app import DATA_DIR
-from app.storage_service import BigQueryService, generate_timestamp, bigquery
+from app.bq_service import BigQueryService, generate_timestamp, bigquery
 from app.email_service import send_email
 
 class NetworkGrapher():
