+++ test/test_bq_grapher.py
@@ -2,7 +2,7 @@ import os
 from networkx import read_gpickle
 
 from app.workers.bq_grapher import NetworkGrapher
-from app.storage_service import BigQueryService
+from app.bq_service import BigQueryService
 
 def test_network_grapher(mock_graph, expected_nodes, expected_edges):
