+++ test/test_pg_grapher.py
@@ -2,7 +2,7 @@ import os
 from networkx import read_gpickle
 
 from app.workers.pg_grapher import NetworkGrapher
-from app.storage_service import BigQueryService
+from app.bq_service import BigQueryService
 
 def test_performance():
